Create CSS folder and one css file
Added a new folder to organize css files for the recipe site.
The lasagna page was the test for the new css file. That is the only
page with a css link. The image was made smaller to fit on the screen.
Unordered lists were made narrow. That is all. :(